requirements.txt
streamlit
fastapi
fastmcp
psycopg2-binary
openai
python-dotenv 